Implement a reverse in-memory LBA (untested) and disable some more extract code. INCOMPLETE, BROKEN

+++ src/containers/segmented_vector.hpp
@@ -9,7 +9,7 @@
 
 #define ELEMENTS_PER_SEGMENT (1 << 14)
 
-template<class element_t, int max_size>
+template<class element_t, size_t max_size>
 class segmented_vector_t
 {
 private:
